Se dejaron en SemenGathering.java, solo 3 campos mandatorios: date, volumne, motility. Y en las demas clases se hicieron los cambios pertientnes para poder permitir estos cambios
git-svn-id: svn://pulsatrix.inbio.ac.cr/atta2/trunk@1430 3aa2e761-397d-4522-80f4-092673de1051

+++ ara-ejb/src/java/org/inbio/ara/eao/germplasm/SemenGatheringEAOLocal.java
@@ -49,7 +49,7 @@ public interface SemenGatheringEAOLocal extends BaseLocalEAO<SemenGathering,Long
 
     public List<Long> findByActiveDoses(Long activeDoses);
 
-    public List<Long> findByStrawSize(Long strawSize);
+    public List<Long> findByStrawSize(Double strawSize);
 
     public List<Long> findBySemenGatheringMethod(Long gatheringSemenMethod);
